requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==5.0.1
djangorestframework==3.15.1
django-heroku==0.3.1
psycopg2==2.9.9
sqlparse==0.4.4
typing_extensions==4.9.0
whitenoise==6.6.0
django_filter
django-cors-headers
django_heroku
